requirements.txt
python-telegram-bot==13.7
flask==2.0.1
flask-socketio==5.1.1
firebase-admin==5.0.1
requests==2.26.0
python-dotenv==0.19.0